refactored dev command
