+++ packages/lodestar-cli/src/commands/dev/index.ts
@@ -1,0 +1,1 @@
+export * from "./command";
